pyproject.toml
[project]
name = "stimm"
version = "0.1.0"
description = "Stimm application with real-time audio processing"
requires-python = ">=3.10"
authors = [
    {name = "Stimm Team", email = "team@stimm.dev"}
]
dependencies = [
    # Core web framework
    "fastapi==0.111.0",
    "uvicorn==0.30.1",
    "pydantic==2.7.4",
    "starlette==0.37.2",
    
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    
    # Vector database and RAG
    "qdrant-client==1.9.0",
    "sentence-transformers==3.0.1",
    "huggingface-hub==0.23.2",
    "rank-bm25==0.2.2",
    
    # AI/ML dependencies
    "torch>=2.1.0",
    "numpy==1.26.4",
    "scipy==1.13.0",
    "onnxruntime>=1.16.0",
    
    # Audio processing
    "sounddevice",
    "soundfile",
    "pyaudio>=0.2.13",
    "webrtcvad==2.0.10",
    "aiortc>=1.8.0",
    "av>=11.0.0",
    
    # LiveKit integration
    "livekit>=0.9.0",
    "livekit-api>=0.9.0",
    "livekit-agents>=0.8.0",
    "livekit-plugins-silero>=0.6.0",
    
    # Providers
    "deepgram-sdk>=3.0.0",
    
    # HTTP and networking
    "aiohttp>=3.10",
    "websockets>=12.0",
    "httpx>=0.27.0",
    
    # Document processing
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.6",
    
    # Environment and utilities
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0", 
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

audio = [
    "pyaudio>=0.2.13",
    "sounddevice",
    "soundfile", 
    "webrtcvad==2.0.10",
]

livekit = [
    "livekit>=0.9.0",
    "livekit-api>=0.9.0", 
    "livekit-agents>=0.8.0",
    "livekit-plugins-silero>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort] 
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=. --cov-report=term-missing"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]

[tool.hatch.envs.default]
skip-install = true
PYTHONPATH = "src"
dependencies = []

[tool.hatch.envs.default.scripts]
# Scripts pour faciliter le lancement
run = "python -m src.main"
debug = "python -m src.main --log-level debug"
test = "pytest tests/ -v"
lint = "black src/ && isort src/ && flake8 src/"

[project.scripts]
stimm-start = "uvicorn.run:app"
